Use GameManager superclass with GameTypes to accommodate additions of multiple games
