+++ src/main/java/dev/deann/Enum/GameType.java
@@ -1,0 +1,5 @@
+package dev.deann.Enum;
+
+public enum GameType {
+    SKYWARS, SURVIVAL_GAMES
+}
